chore: Refactor Cart component to separate Checkout functionality into its own component
